% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_pattern_aru_type}
\alias{add_pattern_aru_type}
\title{Add an ARU to the list of identified ARUs}
\usage{
add_pattern_aru_type(pattern, aru_type)
}
\arguments{
\item{pattern}{regular expression to extract from file path}

\item{aru_type}{Name of ARUtype}
}
\description{
Add an ARU to the list of identified ARUs
}
\examples{

org_pat <- get_pattern("pattern_aru_type")

print(org_pat)

add_pattern_aru_type("CWS\\\\d", "Canadian Wildlife Detector \1")

get_pattern("pattern_aru_type")

set_pattern("pattern_aru_type", org_pat)

}
