% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_wildtrax}
\alias{add_wildtrax}
\title{Add file name formated for Wildtrax to metadata}
\usage{
add_wildtrax(meta)
}
\arguments{
\item{meta}{Data frame. Recording metadata. Output of \code{clean_metadata()}.}
}
\value{
Data frame of metadata with appended column of WildTrax appropriate
file names.
}
\description{
Create and append file name appropriate for uploading data to the Wildtrax
platform \url{https://wildtrax.ca/}.
}
\examples{

m <- clean_metadata(project_files = example_files)
m <- add_wildtrax(m)
m
}
