% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARUtools-package.R
\docType{package}
\name{ARUtools}
\alias{ARUtools-package}
\alias{ARUtools}
\title{ARUtools: Management and Processing of Autonomous Recording Unit (ARU) Data}
\description{
Parse Autonomous Recording Unit (ARU) data and for sub-sampling recordings.
Extract Metadata from your recordings, select a subset of recordings for
interpretation, and prepare files for processing on the
WildTrax \url{https://wildtrax.ca/} platform. Read and process metadata
from recordings collected using the Song Meter and BAR-LT types of ARUs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arutools.github.io/ARUtools/}
  \item \url{https://github.com/ARUtools/ARUtools}
  \item Report bugs at \url{https://github.com/ARUtools/ARUtools/issues}
}

}
\author{
\strong{Maintainer}: David Hope \email{david.hope@ec.gc.ca} (\href{https://orcid.org/0000-0002-2140-4261}{ORCID})

Authors:
\itemize{
  \item Steffi LaZerte \email{sel@steffilazerte.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})
}

Other contributors:
\itemize{
  \item Government of Canada [copyright holder, funder]
}

}
