% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSelPr.R
\name{calc_sel_pr}
\alias{calc_sel_pr}
\title{Calculate Inclusion Probabilities}
\usage{
calc_sel_pr(
  .data,
  ARU_ID_col,
  min_col,
  day_col,
  parms = list(min_range = c(-60, 300), doy_range = c(150, 180), mean_min = 30, sd_min
    = 30, mean_doy = 0, sd_doy = 10, off = 0, log_ = FALSE, fun = "norm")
)
}
\arguments{
\item{.data}{Data frame with Date, times, and ARU IDs}

\item{ARU_ID_col}{ARU id column. Should not be quoted.}

\item{min_col}{Minutes column. Should not be quoted.}

\item{day_col}{Day column. Should not be quoted.}

\item{parms}{list of parameters. See defaults for examples.
 Should include min_range, doy_range, mean_min, sd_min,
mean_doy, sd_doy, off, log_, fun.}
}
\value{
Returns .data with selection probability columns
}
\description{
Calculate Inclusion Probabilities
}
