% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_probabilities.R
\name{calc_selection_probs}
\alias{calc_selection_probs}
\title{Calculate Inclusion Probabilities}
\usage{
calc_selection_probs(
  meta_sun,
  params,
  col_site_id = "site_id",
  col_min = "t2sr",
  col_day = "date"
)
}
\arguments{
\item{meta_sun}{(Spatial) Data frame. Recording meta data with time to
sunrise/sunset. Output of \code{calc_sun()}. Must have at least \code{col_min}, \code{col_day}, and \code{col_site_id}}

\item{params}{list of parameters. See defaults for examples.
Should include min_range, doy_range, mean_min, sd_min,
mean_doy, sd_doy, off, log_, fun.}

\item{col_site_id}{Character. Column name that contains site ids. Deafult \code{site_id}.}

\item{col_min}{Minutes column. Should not be quoted.}

\item{col_day}{Day column. Should not be quoted.}
}
\value{
Returns .data with selection probability columns
}
\description{
Calculate Inclusion Probabilities
}
\examples{
m <- clean_metadata(project_files = example_files)
s <- clean_site_index(example_sites_clean,
                      col_date_time = c("date_time_start", "date_time_end"))
m <- add_sites(m, s)
m <- calc_sun(m)

params <- selection_params()
calc_selection_probs(m, params = params)
}
