% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_dates.R
\name{calculate_sunrise_sunset}
\alias{calculate_sunrise_sunset}
\title{Calculate sunrise sunset}
\usage{
calculate_sunrise_sunset(.data, var_day)
}
\arguments{
\item{.data}{data frame from prep_sunrise_sunset}

\item{var_day}{variable from .data to use as date to calculate sunrise}
}
\value{
Returns a data frame with sunrise and sunset
}
\description{
A wrapper around suncalc::getSunlightTimes to ensure
   timezones and lat/lon are correct.
}
