% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDirectory.R
\name{create_directory_structure}
\alias{create_directory_structure}
\title{Create directory structure for ARU folders}
\usage{
create_directory_structure(hexagons, units, base_dir)
}
\arguments{
\item{hexagons}{Character vector of hexagons or clusters}

\item{units}{Character vector of ARU unit ids. Should include the hexagon in the name}

\item{base_dir}{Base directory to build directory structure from.}
}
\value{
Does not return anything if run. If cancelled returns string notification.
}
\description{
Create directory structure for ARU folders
}
