% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patterns.R
\name{create_pattern_date}
\alias{create_pattern_date}
\title{Create a pattern to match date}
\usage{
create_pattern_date(order = "ymd", sep = c("_", "-"), n_years = 4)
}
\arguments{
\item{order}{Character. Expected order of (y)ear, (m)onth and (d)ate.
Default is "ymd" for Year-Month-Date order.}

\item{sep}{Character vector. Expected (optional) separators between the
values. Can be "" for no separator. Otherwise all separators are optional.}

\item{n_years}{Numeric. Number of digits in Year, either 2 or 4.}
}
\description{
Helper function to create a regular expression pattern to match different
types of date formats in file paths.
}
\examples{
create_pattern_date()
create_pattern_date(sep = "")

}
