% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_dates.R
\name{est_sunrise_sunset}
\alias{est_sunrise_sunset}
\title{Temporary alternative function to calculate sunrise sunset}
\usage{
est_sunrise_sunset(data_in, time_zone)
}
\arguments{
\item{time_zone}{}
}
\description{
Just a placeholder to put into prep_sunrise_sunset potentially
}
