% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_files}
\alias{example_files}
\title{Example recording files}
\format{
\subsection{\code{example_files}}{

A vector with 42 file paths
}
}
\source{
data-raw/data_test.R
}
\usage{
example_files
}
\description{
A vector of examples ARU recording files.
}
\keyword{datasets}
