% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_files_long}
\alias{example_files_long}
\title{Example long-term deployment recording files}
\format{
\subsection{\code{example_files_long}}{

A vector with 614 file paths
}
}
\source{
data-raw/data_long_deployment.R
}
\usage{
example_files_long
}
\description{
A vector of examples ARU recording files. Uses the
\code{example_sites} data, but deploys them for a longer deployment
}
\keyword{datasets}
