% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sites}
\alias{example_sites}
\title{Example site-level meta data}
\format{
\subsection{\code{example_sites}}{

A data frame with 10 rows and 8 columns:
\describe{
\item{Sites}{Site ids}
\item{Date_set_out}{Deployment start date}
\item{Date_removed}{Deployment end date}
\item{ARU}{ARU ids}
\item{lon}{Longitude in decimal degrees}
\item{lat}{Latitude in decimal degrees}
\item{Plots}{Hypothetical extra plot column}
\item{Subplot}{Hypothetical extra subplot column}
}
}
}
\source{
data-raw/data_test.R
}
\usage{
example_sites
}
\description{
A data frame with examples of incorrectly formatted site-level data.
}
\keyword{datasets}
