% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sites_clean}
\alias{example_sites_clean}
\title{Example cleaned site-level meta data}
\format{
\subsection{\code{example_sites_clean}}{

A data frame with 10 rows and 8 columns:
\describe{
\item{site_id}{Site ids}
\item{aru_id}{ARU ids}
\item{date_time_start}{Deployment start date/time}
\item{date_time_end}{Deployment end date/time}
\item{date_start}{Deployment start date}
\item{date_end}{Deployment end date}
\item{longitude}{Latitude in decimal degrees}
\item{latitude}{Longitude in decimal degrees}
}
}
}
\source{
data-raw/data_test.R
}
\usage{
example_sites_clean
}
\description{
A data frame with examples of correctly formatted site-level data.
}
\keyword{datasets}
