% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clip_wav_file.R
\name{format_clip_wave}
\alias{format_clip_wave}
\title{Clip wave to length from given start time in recording}
\usage{
format_clip_wave(
  seg,
  length_clip,
  out_dir,
  sub_dir_col,
  filename_col,
  warn = T
)
}
\arguments{
\item{warn}{}
}
\description{
Clip wave to length from given start time in recording
}
