% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clip_wav_file.R
\name{format_clip_wave_single}
\alias{format_clip_wave_single}
\title{Clip single wave file to length from given start time in recording}
\usage{
format_clip_wave_single(
  in_file,
  out_file,
  length_clip,
  StartTime,
  warn = T,
  l = NULL
)
}
\arguments{
\item{l}{}
}
\description{
Clip single wave file to length from given start time in recording
}
