% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSamples.R
\name{fun_aru_samp}
\alias{fun_aru_samp}
\title{Run the model function}
\usage{
fun_aru_samp(df, N, os, seed, strat_, selprob_id, x, y)
}
\arguments{
\item{df}{Data frame to use with ARU information and selection probabilities}

\item{N}{Number of samples per stratum}

\item{os}{Oversample proportion}

\item{seed}{random number seed}

\item{strat_}{Stratum ID. This is the column grts runs on}

\item{selprob_id}{Selection probability column}

\item{x}{Column in df for x (I've been using doy)}

\item{y}{Column in df for y -time of day or time to sunrise/sunset}
}
\value{
A sampling run from grts
}
\description{
Run the model function
}
