% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSelPr.R
\name{gen_dens_sel_simulation}
\alias{gen_dens_sel_simulation}
\title{Simulate selection probabilities for GRTS}
\usage{
gen_dens_sel_simulation(
  min,
  mean_min,
  sd_min,
  doy,
  mean_doy,
  sd_doy,
  return_dat = F,
  log_ = F,
  fun = "norm",
  off = NULL
)
}
\arguments{
\item{min}{Vector of minutes to sunrise or sunset}

\item{doy}{Vector of minutes to sunrise or sunset (negative is before)}

\item{mean_doy}{Mean of density for days to sun event}

\item{sd_doy}{Standard deviation of days portion of density}

\item{return_dat}{Logical - return just the data or print basic plots}

\item{log_}{Logical - to use log of densities (creates softer patterns)}

\item{fun}{Function to use for densities. Can be "lognorm", "norm", or "cauchy"}

\item{mean_log_min}{Mean (on log scale) of density for minutes to sun event}

\item{sdlog_min}{Standard deviation of minutes portion of density}
}
\value{
Returns either a data frame with selection probabilities or plots of selection probabiliies.
}
\description{
Simulate selection probabilities for GRTS
}
