% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clip_wav_file.R
\name{get_file_info}
\alias{get_file_info}
\title{Get the file info from a recording}
\usage{
get_file_info(file, use_exifr = FALSE)
}
\arguments{
\item{file}{Path to  file}

\item{use_exfir}{Logical use the exifr package, which is slower, but returns more information and allows use of flac files.}
}
\value{
tibble with file information. use_exifr creates data, but takes longer.
}
\description{
Get the file info from a recording
}
\examples{
f <- tempfile()
wav <- download.file("https://www2.cs.uic.edu/~i101/SoundFiles/StarWars3.wav", destfile = f)
get_file_info(f)
get_file_info(f, use_exifr = TRUE)
}
