% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_pattern}
\alias{get_pattern}
\title{Returns the current vector of ARU types}
\usage{
get_pattern(pattern_name)
}
\arguments{
\item{pattern_name}{String of pattern variable to return. One of
"pattern_aru_type", "pattern_check","pattern_data", or "pattern_date_time"}
}
\value{
named character vector
}
\description{
Returns the current vector of ARU types
}
\examples{

get_pattern("pattern_aru_type")
}
