% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wave.R
\name{get_wav_length}
\alias{get_wav_length}
\title{Get the length of a recording in seconds}
\usage{
get_wav_length(path, return_numeric = FALSE)
}
\arguments{
\item{path}{Character. Path to wave file.}

\item{return_numeric}{Logical. Return numeric or character?}
}
\value{
Length of recording in seconds
}
\description{
Get the length of a recording in seconds
}
\examples{
  f <- tempfile()
  w <- tuneR::sine(440, duration = 100000)
  tuneR::writeWave(w, f)
  get_wav_length(f)
}
