% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComplex.R
\name{getvals}
\alias{getvals}
\title{Get accoustic complexity values}
\usage{
getvals(filename, fl, mf = NA, maxf = NA, units_ = "samples")
}
\arguments{
\item{filename}{File name}

\item{fl}{file location}

\item{mf}{minimum frequency for complexity}

\item{maxf}{max freqency for complexity}

\item{units_}{}
}
\value{

}
\description{
Get accoustic complexity values
}
