% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_ARU_type.R
\name{guess_ARU_type}
\alias{guess_ARU_type}
\title{Try to guess the ARU type from a file path}
\usage{
guess_ARU_type(path)
}
\arguments{
\item{path}{Character. Path to wave file}
}
\value{
Tibble with columns 'manufacturer', 'model', and 'aru_type'
}
\description{
Try to guess the ARU type from a file path
}
\examples{

get_pattern("pattern_aru_type")

guess_ARU_type("/path/to/barlt/file.wav")

guess_ARU_type("/path/to/sm/S4A2342.wav")
}
