% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_dates.R
\name{link_gps_locs}
\alias{link_gps_locs}
\title{Link GPS locations with wav files}
\usage{
link_gps_locs(gps_locations, wav_names_log)
}
\arguments{
\item{gps_locations}{gps locatoins}

\item{wav_names_log}{wav list}
}
\value{

}
\description{
Link GPS locations with wav files
}
