% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_dates.R
\name{parse_datetimes_BarLT}
\alias{parse_datetimes_BarLT}
\title{Parse file names of wave files for dates and times.}
\usage{
parse_datetimes_BarLT(
  list_waves,
  site_in_filename,
  site_pattern = "[P|Q]\\\\d+_\\\\d",
  filename_separator = "T|\\\\-|\\\\_|\\\\.",
  tz_loc = "America/Toronto"
)
}
\arguments{
\item{tz_loc}{}
}
\value{

}
\description{
Parse file names of wave files for dates and times.
}
