% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_random_song.R
\name{play_random_track}
\alias{play_random_track}
\title{Play random track}
\usage{
play_random_track(base_folder, random_seed = NULL)
}
\arguments{
\item{base_folder}{Base folder path from which to search from.}
}
\value{
Will not return anything. It does open your media player
}
\description{
Will play a random track from a wave file in a given folder.
}
