% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_for_wind_det.R
\name{prep_for_wind_detection}
\alias{prep_for_wind_detection}
\title{Prepare files for WindNoiseDetection}
\usage{
prep_for_wind_detection(
  wav_files,
  site_pattern,
  output_directory,
  write_to_file = FALSE,
  chunk_size = NULL
)
}
\arguments{
\item{wav_files}{Vector of path to wav files}

\item{site_pattern}{Pattern to extract sites from file names}

\item{output_directory}{Directory path to export files to}

\item{write_to_file}{Logical Should the function write files to output_directory}

\item{chunk_size}{Numeric If not NULL, sets number of files to include in each chunk}
}
\value{
List of vectors for sites, filenames, and filePaths
}
\description{
This function prepares a set of files that can be used with
the fork of WindNoiseDetection found at
https://github.com/dhope/WindNoiseDetection
}
