% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_dates.R
\name{prep_sunrise_sunset}
\alias{prep_sunrise_sunset}
\title{Calculate sunrise and sunset and time to sunset for BarLT recordings}
\usage{
prep_sunrise_sunset(gps_locations, wav_names_log)
}
\arguments{
\item{gps_locations}{data frame generated by process_gps_barlt or process_gps_SM}

\item{wav_names_log}{data frame generated by parse_datetimes}
}
\value{
returns a data frame with time to sunrise
}
\description{
Calculate sunrise and sunset and time to sunset for BarLT recordings
}
