% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_dates.R
\name{prep_sunrise_sunset_BarLT}
\alias{prep_sunrise_sunset_BarLT}
\title{Calculate sunrise and sunset and time to sunset for BarLT recordings}
\usage{
prep_sunrise_sunset_BarLT(gps_locations, wav_names_log)
}
\arguments{
\item{wav_names_log}{}
}
\value{

}
\description{
Calculate sunrise and sunset and time to sunset for BarLT recordings
}
