% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{process_gps_barlt}
\alias{process_gps_barlt}
\title{Process GPS log file for BarLT}
\usage{
process_gps_barlt(base_folder, file_list, deploy_start_date, check_dists, ...)
}
\arguments{
\item{deploy_start_date}{}
}
\value{

}
\description{
Extracts out lat/lon locations from gps log files
  will drop locations before a given date, but does
  not drop duplicate locations
}
