% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{process_log_SM}
\alias{process_log_SM}
\title{Process log files from SongMeters}
\usage{
process_log_SM(folder_base, list_files, site_pattern, return_gps, return_log)
}
\arguments{
\item{folder_base}{Base folder were summary folders locationed}

\item{list_files}{List of files in folder_base}

\item{site_pattern}{site pattern to separate out siteid}

\item{return_gps}{Logical. Should function return GPS locations?}

\item{return_log}{Logical. Should the function return the full log?}
}
\value{
Returns a data frame with lon/lat for each location and date collected
}
\description{
Process log files from SongMeters
}
