% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_log_barlt}
\alias{read_log_barlt}
\title{Parse the log file for BarLT}
\usage{
read_log_barlt(filename)
}
\arguments{
\item{filename}{Character. File location of log file}
}
\value{

}
\description{
Returns the serial number, firware version
and log file parsed as a tibble.
}
