% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_summary_SM4}
\alias{read_summary_SM4}
\title{Parse SM4 logfiles}
\usage{
read_summary_SM4(filename, SiteID_pattern = "SM4A\\\\d{5}")
}
\arguments{
\item{filename}{Character string or vector of character strings of file locations.
Can be a single file or a vector of file locations.}

\item{SiteID_pattern}{Pattern for extracting SiteID from filename. Defaults to 'SM4A\\d{5}',
which is SM4, followed by 5 digits.}
}
\value{
Returns data frame of log files.
}
\description{
Parse SM4 logfiles
}
