% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set_pattern}
\alias{set_pattern}
\title{Set pattern into ARUtools environment}
\usage{
set_pattern(pattern_name, pattern)
}
\arguments{
\item{pattern_name}{string of variable to set}

\item{pattern}{Pattern to add into ARUtools environment}
}
\description{
Set pattern into ARUtools environment
}
\examples{

og_pat <- get_pattern("pattern_date_time")

set_pattern("pattern_date_time", create_pattern_date())

glue::glue("Default pattern: {og_pat}")
glue::glue("Updated pattern: {get_pattern('pattern_date_time')}")

set_pattern("pattern_date_time", og_pat)


}
