% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soxHex.R
\name{soxSpectrogram}
\alias{soxSpectrogram}
\title{Create spectrogram from wave file}
\usage{
soxSpectrogram(
  file.path,
  out.path = "Spectrograms",
  out.app = "test",
  size = list(x = 2000, y = 1000),
  duration = list(start = 0, end = 180),
  sox.file.path
)
}
\arguments{
\item{file.path}{Input must be file with its directory (file.path)}

\item{out.path}{Defult will save in Spectrogram folder in the current working directory otherwise specify location (out.path)}

\item{sox.file.path}{}
}
\value{

}
\description{
Functions from Sam Hache to create spectrograms
}
\details{
Create spectrogram and save it wherever you want
Input must be file with its directory (file.path)
Defult will save in Spectrogram folder in the current working directory otherwise specify location (out.path)
Size specifies the x and y size of the output in pixels must be list
Duration sets the duration of spectrogram default is 3 min and is listed in seconds
}
