% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{task_template}
\alias{task_template}
\title{Example template of tasks for WildTrax}
\format{
\subsection{\code{task_template}}{

A data frame with 14 rows and 13 columns:
\describe{
\item{location}{Site location name}
\item{recording_date_time}{Date time of the recording}
\item{method}{Method of interpretation (generally '1SPT')}
\item{taskLength}{Length of recording in seconds}
\item{transcriber}{Transcriber ID, to be filled in with function}
\item{rain}{Empty character for filling in WildTrax}
\item{wind}{Empty character for filling in WildTrax}
\item{industryNoise}{Empty character for filling in WildTrax}
\item{audioQuality}{Empty character for filling in WildTrax}
\item{taskComments}{Empty character for filling in WildTrax}
\item{internal_task_id}{Empty character for filling in WildTrax}
}
}
}
\source{
data-raw/data_wt_assign_tasks.R
}
\usage{
task_template
}
\description{
A data frame with tasks generated from \code{example_clean} using
the wildRtrax::wt_make_aru_tasks() function. Allows updating of
tasks on WildTrax \url{https://wildtrax.ca/}.
}
\keyword{datasets}
