% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper.R
\name{temp_wavs}
\alias{temp_wavs}
\title{Helper function to create test wave files}
\usage{
temp_wavs(n = 6)
}
\arguments{
\item{n}{Numeric. How many test files to create (up to six). D}
}
\value{
vector of paths to temporary wave files
}
\description{
Creates a directory structure and example wave files in temp folders.
}
\examples{
temp_wavs(n=3)

}
