% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_observers}
\alias{template_observers}
\title{Example template of tasks for WildTrax}
\format{
\subsection{\code{template_observers}}{

A data frame with 4 rows and 2 columns:
\describe{
\item{transcriber}{Interpreter name in Wildtrax system}
\item{hrs}{Number of hours to assign to interpreter}
}
}
}
\source{
data-raw/data_wt_assign_tasks.R
}
\usage{
template_observers
}
\description{
A data frame showing example observers and their effort
}
\keyword{datasets}
