% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_detection_functions.R
\name{wind_detection_summarize_json}
\alias{wind_detection_summarize_json}
\title{Summarize wind detection results}
\usage{
wind_detection_summarize_json(f)
}
\arguments{
\item{f}{filepath for json

#'}
}
\value{
tibble of summarized data from json file
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function takes output from the command line program and summarizes it.
Details of the wind detection software can be found at
\url{https://github.com/dhope/WindNoiseDetection}.
}
\examples{
# example code

 example_json <- system.file("extdata",
 "P71-1__20210606T232500-0400_SS.json",
 package = "ARUtools"
 )

 wind_summary <- wind_detection_summarize_json(example_json)

}
