% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_wildtrax}
\alias{add_wildtrax}
\title{Add Wildtrax file name to metadata}
\usage{
add_wildtrax(meta)
}
\arguments{
\item{meta}{Data frame. Recording metadata. Output of `clean_metadata()`.}
}
\value{
Data frame of metadata with appended column of wildtrax appropriate
  file names.
}
\description{
Create and append file name appropriate for uploading data to the Wildtrax
system.
}
\examples{

m <- clean_metadata(project_files = example_files)
m <- add_wildtrax(m)
m
}
