% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARUtools-package.R
\docType{package}
\name{ARUtools}
\alias{ARUtools}
\alias{ARUtools-package}
\title{ARUtools: CWS Ontario Tools for working with ARU files}
\description{
ARU tools that some might find useful for working with ARU files.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://davidhope.ca/ARUtools/}
}

}
\author{
\strong{Maintainer}: David Hope \email{david.hope@ec.gc.ca} (\href{https://orcid.org/0000-0002-2140-4261}{ORCID})

Authors:
\itemize{
  \item Steffi LaZerte \email{sel@steffilazerte.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})
}

}
