% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSelPr.R
\name{calc_sel_pr}
\alias{calc_sel_pr}
\title{Calculate Inclusion Probabilities}
\usage{
calc_sel_pr(
  meta,
  col_aru_id = "aru_id",
  col_min = "t2sr",
  col_date = "date",
  params = sel_pr_params()
)
}
\arguments{
\item{meta}{Data frame. Recording metadata. Output of \code{clean_metadata()}.}

\item{col_aru_id}{Character. Column name that contains ARU ids. Default \code{aru_id}.}

\item{col_min}{Minutes column. Should not be quoted.}

\item{col_date}{Day column. Should not be quoted.}

\item{params}{list of parameters. See defaults for examples.
Should include min_range, doy_range, mean_min, sd_min,
mean_doy, sd_doy, off, log_, fun.}

\item{.data}{Data frame with Date, times, and ARU IDs}
}
\value{
Returns .data with selection probability columns
}
\description{
Calculate Inclusion Probabilities
}
\examples{
m <- clean_metadata(project_files = example_files)
s <- clean_site_index(example_sites_clean,
                      col_date = c("date_time_start", "date_time_end"))
m <- add_sites(m, s)
m <- calc_sun(m)
calc_sel_pr(m, params = sel_pr_params(doy_range = c(110, 180)))

}
