% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_meta}
\alias{check_meta}
\title{Check output of `clean_metadata()`}
\usage{
check_meta(meta, by_date = FALSE)
}
\arguments{
\item{meta}{Data frame. Recording metadata. Output of `clean_metadata()`.}

\item{by_date}{Logical. Whether to summarize metadata by date. Default
`FALSE`.}
}
\value{
A data frame summarizing the metadata by site_id, aru_type, aru_id,
and (optionally) by date. Presents the number of files, directories, and days
worth of recordings, as well as the minimum and maximum recording times.
}
\description{
Cleaning metadata can take a series of tries. This function helps summarize
and explore the metadata for possible patterns which may help find problems.
}
\examples{
m <- clean_metadata(project_files = example_files)

check_meta(m)
check_meta(m, by_date = TRUE)

}
