% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_problems}
\alias{check_problems}
\title{Check problems in output of `clean_metadata()`}
\usage{
check_problems(
  meta,
  check = c("site_id", "aru_id", "date", "date_time", "longitude", "latitude"),
  path = FALSE,
  by_date = FALSE
)
}
\arguments{
\item{meta}{Data frame. Recording metadata. Output of `clean_metadata()`.}

\item{check}{Character. Character vector of columns to check for missing
values. Default is `site_id`, `aru_id`, `date`, `date_time`, `longitude`
and `latitude`.}

\item{path}{Logical. Whether to return just the file paths which have missing
attributes. Default `FALSE`}

\item{by_date}{Logical.}
}
\value{
A data frame summarizing the metadata by site_id, aru_type, aru_id,
and (optionally) by date. Presents the number of files, directories, and days
worth of recordings, as well as the minimum and maximum recording times.
}
\description{
Cleaning metadata can take a series of tries. This function helps summarize
and explore missing metadata (problems).
}
\examples{

m <- clean_metadata(project_files = example_files, pattern_aru_id = "test")

check_problems(m)
check_problems(m, by_date = TRUE)
check_problems(m, path = TRUE)
}
