% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gps.R
\name{clean_gps}
\alias{clean_gps}
\title{Check and clean GPS data}
\usage{
clean_gps(
  meta,
  gps = NULL,
  dist_cutoff = 100,
  dist_crs = 3161,
  skip_bad = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{meta}{Data frame. Output of `clean_metadata()`.}

\item{gps}{Data frame. GPS data points (optional, but recommended).}

\item{dist_cutoff}{Numeric. Maximum distance (m) between GPS points within a
site. Default is 100m but can be set to `Inf` to skip.}

\item{dist_crs}{Numeric. Coordinate Reference System to use when calculating
distance (should be one with m).}

\item{verbose}{Logical.}
}
\description{
Check and clean GPS data either from ARU logs or user-supplied GPS data
(preferred). GPS points are checked for obvious problems (expected range,
distance cutoffs and timing) then attached to the meta data frame.
}
