% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Clean aru metadata}
\usage{
clean_metadata(type, folder_base, gps_locations = NULL, ...)
}
\arguments{
\item{type}{BarLT or SM4}

\item{folder_base}{Folder base directory}

\item{gps_locations}{GPS locations in lat/lon format with SiteID
Must also include timezone (tz) and deploy date.}
}
\value{

}
\description{
Clean aru metadata
}
