% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Clean aru metadata}
\usage{
clean_metadata(type, folder_base, sitename)
}
\arguments{
\item{type}{BarLT or SM4}

\item{folder_base}{Folder base directory}

\item{sitename}{Site name}
}
\value{

}
\description{
Clean aru metadata
}
