% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Clean aru metadata}
\usage{
clean_metadata(
  type,
  folder_base,
  included_subfolders = "all",
  gps_locations = NULL,
  file_split_pattern = "T|\\\\-|\\\\_|\\\\.",
  site_pattern = "[P|Q]\\\\d+_\\\\d",
  ...
)
}
\arguments{
\item{type}{BarLT or SM4}

\item{folder_base}{Folder base directory}

\item{included_subfolders}{Defaults to 'all', which means all subfolders included.
To include only some folders (at first level only),
provide a vector with folder names to include.}

\item{gps_locations}{GPS locations in lat/lon format with SiteID
Must also include timezone (tz) and deploy date.}

\item{file_split_pattern}{Regular expression to split sound file name into pieces.
May need to adjust if non-standard naming convention is used.}

\item{site_pattern}{Regular expression to extract site name from folder path.}

\item{...}{Extra parameters for devleopment. Currently works with:
- file_ext: default is '.wav'
- site_in_filename: Logical: default is FALSE
- deploy_start_date:  default is 2 days before first recording (dmy format)
- check_dists: Logical to check distances from GPS within site. Default is TRUE
- dist_cutoff: Cutoff to raise error in distance check. Default is 100 (in meters).}
}
\description{
Clean aru metadata
}
