% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Clean aru metadata}
\usage{
clean_metadata(
  project_dir = NULL,
  project_files = NULL,
  file_type = "wav",
  subset = NULL,
  subset_type = "keep",
  pattern_site = "[P|Q]\\\\d+(_|-)\\\\d",
  pattern_aru_id = create_pattern_aru_id(),
  pattern_date = create_pattern_date(),
  pattern_time = create_pattern_time(),
  pattern_dt_sep = create_pattern_dt_sep(),
  order_date = "ymd",
  quiet = FALSE
)
}
\arguments{
\item{type}{BarLT or SM4}

\item{folder_base}{Folder base directory}

\item{included_subfolders}{Defaults to 'all', which means all subfolders included.
To include only some folders (at first level only),
provide a vector with folder names to include.}

\item{gps_locations}{GPS locations in lat/lon format with SiteID
Must also include timezone (tz) and deploy date.}

\item{file_split_pattern}{Regular expression to split sound file name into pieces.
May need to adjust if non-standard naming convention is used.}

\item{site_pattern}{Regular expression to extract site name from folder path.}

\item{...}{Extra parameters for devleopment. Currently works with:
- file_ext: default is '.wav'
- site_in_filename: Logical: default is FALSE
- deploy_start_date:  default is 2 days before first recording (dmy format)
- check_dists: Logical to check distances from GPS within site. Default is TRUE
- dist_cutoff: Cutoff to raise error in distance check. Default is 100 (in meters).}
}
\description{
Clean aru metadata
}
\examples{

clean_metadata(subset = "P71")
clean_metadata()

}
