% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Extract and clean ARU metadata from file names}
\usage{
clean_metadata(
  project_dir = NULL,
  project_files = NULL,
  file_type = "wav",
  subset = NULL,
  subset_type = "keep",
  pattern_site_id = create_pattern_site_id(),
  pattern_aru_id = create_pattern_aru_id(),
  pattern_date = create_pattern_date(),
  pattern_time = create_pattern_time(),
  pattern_dt_sep = create_pattern_dt_sep(),
  order_date = "ymd",
  quiet = FALSE
)
}
\arguments{
\item{project_dir}{Character. Directory where project files are stored. File
paths will be used to extract information Must have one of `project_dir` or
`project_files`.}

\item{project_files}{Character. Vector of project file paths. Must have one
of `project_dir` or `project_files`.}

\item{file_type}{Character. Type of file (extension) to summarize. Default wav.}

\item{subset}{Character. Text pattern to mark a subset of files/directories
to either `keep` or `omit` (see `subset_type`)}

\item{subset_type}{Character. Either `keep` (default) or `omit`
files/directories which match the pattern in `subset`.}

\item{pattern_site_id}{Character. Regular expression to extract site ids. See
`create_pattern_site_id()`.}

\item{pattern_aru_id}{Character. Regular expression to extract ARU ids. See
`create_pattern_aru_id()`.}

\item{pattern_date}{Character. Regular expression to extract dates. See
`create_pattern_date()`.}

\item{pattern_time}{Character. Regular expression to extract times. See
`create_pattern_time()`.}

\item{pattern_dt_sep}{Character. Regular expression to mark separators betwen
dates and times. See `create_pattern_dt_sep()`.}

\item{order_date}{Character. Order that the date appears in. One of "ymd"
(default), "mdy", "dmy".}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\value{
Data frame with extracted metadata
}
\description{
Using regular expressions, metadata is extracted from file names and
directory structure, checked and cleaned.
}
\details{
See `vignette("customizing", package = "ARUtools")` for details on customizing
`clean_metadata()` for your project.
}
\examples{
clean_metadata(project_files = example_files)
clean_metadata(project_files = example_files, subset = "P02")

}
