% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sites.R
\name{clean_site_index}
\alias{clean_site_index}
\title{Title}
\usage{
clean_site_index(
  site_index,
  col_aru_id = "aru_id",
  col_site_id = "site_id",
  col_date_time = "date",
  col_coords = c("longitude", "latitude"),
  col_extra = NULL,
  resolve_overlaps = TRUE
)
}
\arguments{
\item{col_coords}{}
}
\description{
Title
}
