% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{meta}{Data frame. Recording metadata. Output of `clean_metadata()`.}

\item{meta_gps}{Data frame. Recording metadata with added coordinates.
Output of `clean_metadata()` and then `add_sites()`.}
}
\description{
Common arguments and documentation for various functions
}
\details{
Use `@inheritParams common_docs` to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
