% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{project_dir}{Character. Directory where project files are stored. File
paths will be used to extract information and must actually exist.}

\item{project_files}{Character. Vector of project file paths. These paths can
be absolute or relative to the working directory, and don't actually need
to point to existing files unless you plan to use `clean_gps()` or other
sampling steps down the line. Must be provided if `project_dir` is `NULL`.}

\item{subset}{Character. Text pattern to mark a subset of files/directories
to either `keep` or `omit` (see `subset_type`)}

\item{subset_type}{Character. Either `keep` (default) or `omit`
files/directories which match the pattern in `subset`.}

\item{meta}{Data frame. Recording metadata. Output of `clean_metadata()`.}

\item{meta_sites}{(Spatial) Data frame. Recording metadata with added
coordinates. Output of `clean_metadata()` and then `add_sites()` (with
either `clean_gps()` or `clean_site_index()`).}

\item{date}{Logical. Whether to summarize output by date (as well as
`site_id` and `aru_id`. Default `FALSE`.}
}
\description{
Common arguments and documentation for various functions
}
\details{
Use `@inheritParams common_docs` to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
