% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{count_files}
\alias{count_files}
\title{Count files in a project directory}
\usage{
count_files(project_dir, subset = NULL, subset_type = "keep")
}
\arguments{
\item{project_dir}{Character. Directory where project files are stored. File
paths will be used to extract information Must have one of `project_dir` or
`project_files`.}

\item{subset}{Character. Text pattern to mark a subset of files/directories
to either `keep` or `omit` (see `subset_type`)}

\item{subset_type}{Character. Either `keep` (default) or `omit`
files/directories which match the pattern in `subset`.}
}
\description{
Helper function to explore the number of files in a directory, recursively.
}
