% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directory structure for ARU folders}
\usage{
create_dirs(
  hexagons,
  aru_ids,
  base_dir = NULL,
  dir_list = FALSE,
  dry_run = TRUE
)
}
\arguments{
\item{hexagons}{Character vector. Hexagon or cluster names for folder names.}

\item{aru_ids}{Character vector. ARU unit ids. Should include the
hexagon/cluster id in the name.}

\item{base_dir}{Character. Base directory to build directory structure in.}
}
\value{
Does not return anything if run. If cancelled returns string notification.
}
\description{
Create directory structure for ARU folders
}
\examples{
# Default is to do a dry-run (don't actually create the directories)
create_dirs(hexagons = c("site1", "site2", "site3"),
            aru_ids = c("site1_sm01", "site1_sm02", "site2_sm03", "site2_sm04",
                        "site3_sm05", "site3_sm06"),
            base_dir = "Recordings")

# Get a list of directories which would be created
create_dirs(hexagons = c("site1", "site2", "site3"),
            aru_ids = c("site1_sm01", "site1_sm02", "site2_sm03", "site2_sm04",
                        "site3_sm05", "site3_sm06"),
            base_dir = "Recordings", dir_list = TRUE)

\dontrun{
# Create directories AND return a list of those created
d <- create_dirs(hexagons = c("site1", "site2", "site3"),
                 aru_ids = c("site1_sm01", "site1_sm02", "site2_sm03", "site2_sm04",
                             "site3_sm05", "site3_sm06"),
                 base_dir = "Recordings", dir_list = TRUE,
                 dry_run = FALSE)
d
}
}
