% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_clean}
\alias{example_clean}
\title{Example cleaned recording meta data}
\format{
\subsection{\code{example_clean}}{

A data frame with 42 rows and 10 columns:
\describe{
\item{file_name}{Name of the file}
\item{type}{File type}
\item{path}{Relative file path including file name}
\item{aru_type}{ARU model}
\item{aru_id}{ARU ids}
\item{site_id}{Site ids}
\item{date_time}{Recording date/time}
\item{date}{Recording date}
\item{longitude}{Latitude in decimal degrees}
\item{latitude}{Longitude in decimal degrees}
}
}

## `task_template`
A data frame with 14 rows and 13 columns:
\describe{
  \item{location}{Site location name}
  \item{recording_date_time}{Date time of the recording}
  \item{method}{Method of interpretation (generally '1SPT')}
  \item{taskLength}{Length of recording in seconds}
  \item{transcriber}{Transcriber ID, to be filled in with function}
  \item{rain}{Empty character for filling in WildTrax}
  \item{wind}{Empty character for filling in WildTrax}
  \item{industryNoise}{Empty character for filling in WildTrax}
  \item{audioQuality}{Empty character for filling in WildTrax}
  \item{taskComments}{Empty character for filling in WildTrax}
  \item{internal_task_id}{Empty character for filling in WildTrax}
}
}
\source{
data-raw/data_test.R

data-raw/task_template.R
}
\usage{
example_clean

example_clean
}
\description{
A data frame with examples of correctly formatted metadata with added
site-level information

A data frame with tasks generated from `example_clean` using
the wildRtrax::wt_make_aru_tasks() function
}
\keyword{datasets}
