% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSelPr.R
\name{gen_dens_sel_simulation}
\alias{gen_dens_sel_simulation}
\title{Simulate selection probabilities for GRTS}
\usage{
gen_dens_sel_simulation(
  parms = ARUtools::default_selection_parameters,
  selection_variable = psel_normalized,
  return_dat = F,
  ...
)
}
\arguments{
\item{parms}{List of parameters for simulation see ?ARUtools::default_selection_parameters for names and default values}

\item{selection_variable}{unquoted variable name of which options are psel, psel_doy, psel_tod, psel_std, psel_scaled, or psel_normalized. Default is psel_normalized}

\item{return_dat}{Logical - return just the data or print basic plot}
}
\value{
Returns either a data frame with selection probabilities or plots of selection probabiliies.
}
\description{
Simulate selection probabilities for GRTS
}
\examples{
# gen_dens_sel_simulation(min =  -70:240,doy = 121:201,
#               parms = ARUtools::default_selection_parameters,
#             selection_variable = psel_normalized , return_dat=F)
}
