% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clip_wav_file.R
\name{get_wav_length}
\alias{get_wav_length}
\title{Get the length of a recording in seconds}
\usage{
get_wav_length(wave_file, return_numeric = FALSE)
}
\arguments{
\item{wave_file}{Character. Path to wave file}

\item{return_numeric}{Logical. Return numeric or character?}
}
\value{
Length of recording in seconds
}
\description{
Get the length of a recording in seconds
}
\examples{
f <- tempfile()
wav <- download.file("https://www2.cs.uic.edu/~i101/SoundFiles/StarWars3.wav", destfile = f)
get_wav_length(f)
}
