% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_dates.R
\name{parse_datetimes}
\alias{parse_datetimes}
\title{Parse file names of wave files for dates and times.}
\usage{
parse_datetimes(wav_names_log, tz_loc = "America/Toronto")
}
\arguments{
\item{wav_names_log}{data frame passed from clean_metadata. must include yyyymmdd and hhmmss}

\item{tz_loc}{Time zone for location. Default to "America/Toronto"}
}
\value{
Returns a data frame with filenames parsed to date & time.
}
\description{
Parse file names of wave files for dates and times.
}
