% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_random_song.R
\name{play_random_track}
\alias{play_random_track}
\title{Play random track}
\usage{
play_random_track(base_folder, file_list = NULL, random_seed = NULL)
}
\arguments{
\item{base_folder}{Base folder path from which to search from.}

\item{file_list}{Vector of strings with file locations. If it is null, will search for them manually. Should be relative to base_folder}
}
\value{
Will not return anything. It does open your media player
}
\description{
Will play a random track from a wave file in a given folder.
}
