% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{process_gps_SM}
\alias{process_gps_SM}
\title{Process GPS locations for SongMeters}
\usage{
process_gps_SM(folder_base, list_files)
}
\arguments{
\item{folder_base}{Base folder were summary folders locationed}

\item{list_files}{List of files in folder_base}
}
\value{
Returns a data frame with lat/lon for each location and date collected
}
\description{
Process GPS locations for SongMeters
}
