% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_and_check_barlt_gps}
\alias{read_and_check_barlt_gps}
\title{Read barlt gps and check locations}
\usage{
read_and_check_barlt_gps(.x)
}
\arguments{
\item{.x}{Path to gps file from barLT}
}
\value{
Returns paths chosen by user
}
\description{
Read barlt gps and check locations
}
