% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_detection_functions.R
\name{wind_detection_pre_processing}
\alias{wind_detection_pre_processing}
\title{Pre-processing of files for Wind Detection program}
\usage{
wind_detection_pre_processing(
  wav_vector,
  site_pattern,
  chunksize,
  file_drive,
  output_directory = NULL,
  write_output = FALSE
)
}
\arguments{
\item{write_output}{}
}
\description{
Pre-processing of files for Wind Detection program
}
