% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_detection_functions.R
\name{wind_detection_pre_processing}
\alias{wind_detection_pre_processing}
\title{Pre-processing of files for Wind Detection program}
\usage{
wind_detection_pre_processing(
  wav_vector,
  site_pattern,
  chunksize,
  file_drive,
  output_directory = NULL,
  write_output = FALSE
)
}
\arguments{
\item{wav_vector}{vector of wave files}

\item{site_pattern}{regular expression to extract site names}

\item{chunksize}{numeric size of chunks to split wav_vector into}

\item{file_drive}{Path to replace for use of cygwin}

\item{output_directory}{string - output directory to write files to}

\item{write_output}{logical - if should write output to files.}
}
\value{
List including filePath, filenames, and sites suitable for wind software.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function takes a vector of wave file names and returns a list
of three vectors that can be provided to the wind detection software or
written to files that the software can read. Details of the wind detection
software can be found at
\url{https://github.com/dhope/WindNoiseDetection}
}
