% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_detection_functions.R
\name{wind_detection_summarize_json}
\alias{wind_detection_summarize_json}
\title{Summarize wind detection results}
\usage{
wind_detection_summarize_json(f, json_string_regex = "/[\\\\w|\\\\d|_|-]+")
}
\arguments{
\item{f}{filepath for json}

\item{json_string_regex}{string. Regex to extract from file path}
}
\value{
tibble of summarized data from json file
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function takes output from the command line program and summarizes it.
Details of the wind detection software can be found at
\url{https://github.com/dhope/WindNoiseDetection}.
}
