% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_detection_functions.R
\name{wind_detection_summarize_json}
\alias{wind_detection_summarize_json}
\title{Summarize wind detection results}
\usage{
wind_detection_summarize_json(f, json_string_regex = "/[\\\\w|\\\\d|_|-]+")
}
\arguments{
\item{f}{filepath for json}

\item{json_string_regex}{string. Regex to extract from file path}
}
\description{
Summarize wind detection results
}
