% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_assign_tasks.R
\name{wt_assign_tasks}
\alias{wt_assign_tasks}
\title{Assign tasks for wildtrax interpretation}
\usage{
wt_assign_tasks(
  wt_task_template_in,
  interp_hours_file,
  wt_task_output_file,
  interp_hours_column
)
}
\arguments{
\item{wt_task_template_in}{Path to csv template downloaded from WilldTrax website listing all tasks}

\item{interp_hours_file}{Path to number of hours for each interpreter. Must be csv and include}

\item{wt_task_output_file}{Path to csv of output file for uploading to wildtrax. If left as NULL will not write file}

\item{interp_hours_column}{LazyEval column name with hours for interpreters}
}
\value{
Returns a list with a tibble of assigned tasks and a summary tibble.
}
\description{
Assign tasks for wildtrax interpretation
}
